+++ thoth-export-server/src/specification_query.rs
@@ -1,3 +1,4 @@
+use futures::{stream, StreamExt, TryStreamExt};
 use std::convert::{TryFrom, TryInto};
 use std::sync::Arc;
 use thoth_client::{QueryParameters, ThothClient, Work};
@@ -5,6 +6,9 @@ use thoth_errors::{ThothError, ThothResult};
 use uuid::Uuid;
 
 use crate::record::MetadataSpecification;
+
+const CONCURRENT_REQUESTS: usize = 2;
+const PAGINATION_LIMIT: i64 = 100;
 
 enum SpecificationRequest {
     ByWork,
@@ -41,9 +45,44 @@ impl SpecificationQuery {
     pub(crate) async fn by_publisher(self, publisher_id: Uuid) -> ThothResult<Vec<Work>> {
         let parameters: QueryParameters =
             QueryConfiguration::by_publisher(self.specification).try_into()?;
-        self.thoth_client
-            .get_works(Some(vec![publisher_id]), parameters)
-            .await
+
+        // get the total work count to figure out how to paginate the results
+        let work_count = self
+            .thoth_client
+            .get_work_count(Some(vec![publisher_id]))
+            .await?;
+        let total_pages = (work_count / PAGINATION_LIMIT) + 1;
+        // get a vector of all page offsets we will need
+        let offsets = match total_pages {
+            1 => vec![0], // otherwise a range of (1..1) gives us nothing
+            _ => (1..total_pages)
+                .map(|current_page| (current_page - 1) * PAGINATION_LIMIT)
+                .collect::<Vec<i64>>(),
+        };
+
+        // make concurrent requests iterating the list of offsets to asynchronously obtain all pages
+        let mut works_pages = stream::iter(offsets)
+            .map(|offset| {
+                let client = &self.thoth_client;
+                async move {
+                    client
+                        .get_works(
+                            Some(vec![publisher_id]),
+                            PAGINATION_LIMIT,
+                            offset,
+                            parameters,
+                        )
+                        .await
+                }
+            })
+            .buffer_unordered(CONCURRENT_REQUESTS);
+
+        // merge all pages
+        let mut works: Vec<Work> = vec![];
+        while let Some(page) = works_pages.try_next().await? {
+            works.extend(page);
+        }
+        Ok(works)
     }
 }