+++ CHANGELOG.md
@@ -6,6 +6,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [[0.9.12]](https://github.com/thoth-pub/thoth/releases/tag/v0.9.12) - 2023-02-17
+### Changed
+  - Reduce number of concurrent requests
+
+## [[0.9.11]](https://github.com/thoth-pub/thoth/releases/tag/v0.9.11) - 2023-02-17
+### Changed
+  - Upgrade rust to `1.67.1` in production and development `Dockerfile`
+  - Upgrade build dependencies (npm `v9.5.0`, node `v18.14.1` and rollup `v3.15.0`) in production and development `Dockerfile`
+
+## [[0.9.10]](https://github.com/thoth-pub/thoth/releases/tag/v0.9.10) - 2023-02-17
+### Changed
+  - Include `limit` and `offset` in `thoth-client`'s works query
+  - Paginate `get_works` requests in export API using concurrent requests
+  - Input number of actix workers via CLI arguments
+
+### Added
+  - Work count query to `thoth-client`
+
 ## [[0.9.9]](https://github.com/thoth-pub/thoth/releases/tag/v0.9.9) - 2023-02-16
 ### Changed
   - Upgrade `actix-web` to v4.3.0